+++ src/main/java/racing/view/InputView.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
 
-import racing.CarMoveCountDto;
-import racing.CarNamesDto;
+import racing.view.dto.CarMoveCountDto;
+import racing.view.dto.CarNamesDto;
 
 public class InputView {
 	private static final String CAR_NAME_QUESTION = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
